requirements.txt
streamlit
pandas
numpy
faiss-cpu
pillow
tqdm
python-dotenv
openai
langchain-core
langchain-openai
